requirements.txt
django
djangorestframework
python-dotenv
requests
gunicorn
whitenoise
dj-database-url
psycopg2-binary
pillow
moviepy